Moving stdin reading code to more generic package
