+++ src/cli/BUILD
@@ -29,3 +29,13 @@ go_test(
         '//third_party/go:testify',
     ],
 )
+
+go_test(
+    name = 'stdin_test',
+    srcs = ['stdin_test.go'],
+    data = ['test_data'],
+    deps = [
+        ':cli',
+        '//third_party/go:testify',
+    ],
+)
