+++ src/cli/test_data/stdin.txt
