+++ src/utils/BUILD
@@ -23,13 +23,3 @@ go_bindata(
     ],
     prefix = 'tools/misc',
 )
-
-go_test(
-    name = 'utils_test',
-    srcs = ['utils_test.go'],
-    data = ['test_data/stdin.txt'],
-    deps = [
-        ':utils',
-        '//third_party/go:testify',
-    ],
-)
